add layout page to reuse and add index.js in components
