+++ src/components/index.js
@@ -1,0 +1,10 @@
+export { default as Navbar } from './Navbar';
+export { default as Footer } from './Footer';
+export { default as LayoutPage } from './LayoutPage';
+export { default as RoomsGrid } from './RoomsGrid';
+export { default as RoomsCard } from './RoomCard';
+export { default as BookForm } from './BookForm';
+export { default as ServicesGrid } from './ServicesGrid';
+export { default as ServiceCard } from './ServiceCard';
+export { default as ImageSlider } from './ImageSlider';
+export { default as Dropdown } from './Dropdown';
