+++ src/pages/RoomDetail.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const RoomDetail = () => {
+  return (
+    <div>RoomDetail</div>
+  )
+}
+
+export default RoomDetail
