+++ src/pages/ServicePage.jsx
@@ -1,45 +1,13 @@
-import React, { useState } from "react";
-import { LoginPopupModal, RegisterPopupModal } from "../auth/forms/";
-import Navbar from "../components/Navbar";
-import Footer from "../components/Footer";
-import ServicesGrid from "../components/ServicesGrid";
+import React from "react";
 
-const ServicePage = () => {
-  const [loginPopup, setLoginPopup] = useState(false);
-  const [registerPopup, setRegisterPopup] = useState(false);
+import { LayoutPage, ServicesGrid } from "../components";
 
-  const toggleLoginPopup = () => {
-    setLoginPopup(!loginPopup);
-  };
-
-  const toggleRegisterPopup = () => {
-    setRegisterPopup(!registerPopup);
-  };
-
-  const switchRegisterAndLogin = () => {
-    toggleLoginPopup();
-    toggleRegisterPopup();
-  };
-
+const BookingRoom = () => {
   return (
-    <div className="flex flex-col h-screen">
-      <div className="z-20">
-        <Navbar setLoginPopup={setLoginPopup} setRegisterPopup={setRegisterPopup} />
-      </div>
-
-      {/* <ServiceCard /> */}
+    <LayoutPage>
       <ServicesGrid />
-
-      {/* Render login/register form */}
-      {loginPopup && <LoginPopupModal toggleLoginPopup={toggleLoginPopup} switchRegister={switchRegisterAndLogin} />}
-
-      {registerPopup && (
-        <RegisterPopupModal toggleRegisterPopup={toggleRegisterPopup} switchLogin={switchRegisterAndLogin} />
-      )}
-
-      <Footer />
-    </div>
+    </LayoutPage>
   );
 };
 
-export default ServicePage;
+export default BookingRoom;
